requirements.txt
pandas
requests
chardet
gspread
google-auth